Fix type

+++ python_variable_annotations/3-to_str.py
@@ -2,6 +2,6 @@
 """Convert an integer to a string. function: to_str(n: int) -> str"""
 
 
-def to_str(n: int) -> str:
+def to_str(n: float) -> str:
     """Convert an integer to a string. function to_str(n: int) -> str"""
     return str(n)
